+++ CHANGES.rst
@@ -3,6 +3,10 @@
 
 - Added Password Conversion Support for Protected Documents.
 
+- Added docstrings to the XML-RPC methods, which should enable the
+  introspection functions to be useful.
+
+- Closed #181: --convert-to was ignored in some cases.
 
 3.4 (2025-09-26)
 ----------------